feat: реализовала веши под базу данных

+++ src/main/java/ru/practicum/shareit/item/storage/ItemStorage.java
@@ -5,11 +5,11 @@ import ru.practicum.shareit.item.model.Item;
 import java.util.List;
 
 public interface ItemStorage {
-    List<Item> getItemsByUser(Integer userId);
+    List<Item> getItemsByUser(Long userId);
 
     List<Item> getItemByText(String text);
 
-    Item getItemById(Integer itemId);
+    Item getItemById(Long itemId);
 
     Item addItem(Item item);
